I18N: GnucashGenerInvoiceEntry.Action now is not locale-specific any more.
Note: The author chose not to use the mechanism for i18n of messages (with messages_<lang>.properties
files), because a) this would necessitate the inclusion of more libraries, b) appears oversized
for this. In fact, it's not a "message" in the narrower sense, but a semi-internal string that
we are localizing here (a code-comment-hybrid, so to speak).