Module "Base":  - Improved / re-iterated class FixedPointNumber; there were    some inconsistencies that I have ironed out now.  - Improved test coverage of that class.
This lead to changes in other modules and possibly code from people
who use this lib (so, it's not backward-compatible). I am very well
aware of the implications of this (and this is one of the reasons
why I hesitated so long before implementing the changes), but I think
that the benefits outweigh the drawbacks mid-term (we had backward-
incompatibilities in the last few releases for other good reasons
anyway...).

I am not convinced, however, that the current state is the best
solution in the long run -- it's just "OK for a while" and subject
to a major re-evaluation, possibly in a future major release.
(Cf. notes in README for module "Base".)