Module "Tools": GetCmdtyList: Inserted code for mode "ISIN"
Note that, as opposed to securities in KMyMoney, commodities in GnuCash
do not have a type (at least, not in the same sense -- some consider
the name space to be sort of a type, but that is a different story
here).

Cf sister project, commit 0b4b5e9dd8a3b37aeb82e6780cb2eab7cb5f74c4.