Module "Base": Introduced set of new (dummy) ID types: GCshAcctID, GCshTrxID, etc. (all inherit from GCshID).
These types will be introduced in the other modules, esp. "API" and "API Extensions".

We do this in order to leverage Java's static type safety; certain errors
will become pracitally impossible, both for developer(s) of this lib and its
users. (I already found such an error when making the changes -- it was
in the test cases.)

Another aspect is better symmetry with the sister project, although this
is of minor importance here (in the sister project, we have truly different
IDs for accounts, transactions, etc., whereas here, all IDs are essentially
a UUID). But still...